Date formats: restrict formats of dates
Invalid date formats are accepted.

Invalid date formats can cause the tasks set by users to be nonsensical.
It will also invalidate the future implementation of listing tasks with
incoming due dates.

Let's
* prevent invalid date formats from being accepted
* allow conversion of date formats

This allows dates and times in tasks to be sensible, as well as maintain
a default date format.