Manager classes: modify for new Command classes
Redistribute functions in Bob, Parser and UiManager to account for
the new Command classes.

This makes the code in each class more concise and relevant to their
purpose.

Let's
* Move scanner from Bob to UiManager
* Move executeCommand to Parser, with new name parseCommand

This narrows down the scope of functions and the scanner to only
the scopes they are relevant in. Thus, code quality improves.