Fix deep nesting within certain classes
Deep nesting occurs in some functions, through a combination of
for loops, while loops and if statements.

Deep nesting causes code within these nestings to be hard to track,
since they have many restrictions in different statements.

Let's ensure that there is no more than 2 nesting in all of our
functions.

Limiting deep nesting to a depth of 2 allows some flexibility with
if statements in for or while loops, but avoids confusion with any
further nesting.