Use the default branch
Some if statements do not have an else statement when they should
be checking if invalid conditions are fulfilled.

If we simply assume these invalid conditions will not be reached,
it will cause problems if they are reached, which will cause bugs.

Let's add else clauses to if statements and default branches to
switch cases.