Add restraint to createCommand for events
When creating events, the end date can be before or at the same
time as the start date.

This is unsensible, and so should be rejected, with a message
telling the user that this they may have made a mistake.

Let's
* Implement a function in DateManager to enforce this restraint
* Add test functions in DateManagerTest to test this restraint

Since we are just handling potential unwanted commands, we should
create test functions after implementing the restraint functions
to test if they work as intended.