Add restraint to createCommand for same tasks
Tasks that already exist in the task list can be added again.

This is unsensible, and so should be rejected, with a message
telling the user that this they may have made a mistake.

Let's
* Implement a function in TaskManager to enforce this restraint
* Add test functions in TaskManagerTest to test this restraint

Since we are just handling potential unwanted commands, we should
create test functions after implementing the restraint functions
to test if they work as intended.