Update label in user dialog box
The text in the user's DialogBox is left aligned, which may not be
user-friendly.

Since the user icon is on the right, the text should be right-aligned
to enhance user experience for the UI.

Let's change the alignment of the label to be right-aligned in
getUserDialog().

Changing the alignment in the function itself allows manual alignment
based on which character is 'speaking'.