Fix GUI bug
GUI is unable to be displayed due to incorrect path name.

Without the GUI being displayed, users cannot enter commands and
utilize the program.

Let's update the path name in MainWindows.fxml to the correct
path name.

Doing so, allows the GUI to display, and so the program to be able
to be utilized.