Update build.gradle and src classes to support cross-platform
The program can only run on Windows.

It is desirable for users to be able to run the program regardless
of the OS they use.

Let's update build.gradle and src classes classes to support
cross-platform functionality.

Updating classes package-by-package allows concise testing of the
individual classes, to ensure modifications do not cause bugs in
the program.