Update error message for CreateCommand
Error message for when dates are not given to create deadlines or
events does not mention time is allowed.

Users may not know they can input time, which may worsen user
experience.

Let's modify the error message to indicate time is also allowed.

Modifying the error message instead of displaying it to the user
outright will prevent information overload as soon as the GUI starts
up.